Separated implemention in clgetValp.cc+addeding building test/tst{2,3}.cc executables.
